+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Core
 
 * Remove unnecessary warnings when Pygmo is not installed.
   (See `!286 <https://gitlab.com/esa/pyxel/-/merge_requests/286>`_).
+* Remove parallel computing with Numba.
+  (See `!290 <https://gitlab.com/esa/pyxel/-/merge_requests/290>`_).
 
 Documentation
 ~~~~~~~~~~~~~